# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  app.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
  handler.ApplyCouponReq:
    properties:
      coupon_id:
        type: integer
    type: object
...
  handler.OrderListResponse:
    properties:
      items:
        items:
          $ref: '#/definitions/model.Order'
        type: array
      page:
        type: integer
      size:
        type: integer
      total:
        type: integer
    type: object
  handler.OrderWithPaymentResponse:
    properties:
      coupon:
        $ref: '#/definitions/service.MyCoupon'
      order:
        $ref: '#/definitions/model.Order'
      payment:
        $ref: '#/definitions/handler.PaymentResponse'
    type: object
  handler.PaymentCallbackReq:
    properties:
      notify_data:
        type: string
      payment_id:
        type: string
      status:
        description: paid / failed / refunded
        type: string
    required:
    - payment_id
    - status
    type: object
  handler.PaymentResponse:
    properties:
      CreatedAt:
        type: string
      ID:
        type: integer
      UpdatedAt:
        type: string
      amount_cents:
        type: integer
      notify_data:
        type: string
      order_id:
        type: integer
      payment_id:
        type: string
      status:
        $ref: '#/definitions/model.PaymentStatus'
    type: object
  handler.PollOrderResponse:
    properties:
      message:
        type: string
      order:
        $ref: '#/definitions/service.OrderWithPayment'
      order_num:
        type: string
      payment_id:
        type: string
      status:
        type: string
    type: object
...
  model.CouponStatus:
    enum:
    - available
    - used
    - expired
    type: string
    x-enum-varnames:
    - CouponStatusAvailable
    - CouponStatusUsed
    - CouponStatusExpired
  model.CouponType:
    enum:
    - full_cut
    - discount
    type: string
    x-enum-comments:
      CouponTypeDiscount: 折扣
      CouponTypeFullCut: 满减
    x-enum-descriptions:
    - 满减
    - 折扣
    x-enum-varnames:
    - CouponTypeFullCut
    - CouponTypeDiscount
  model.Order:
    properties:
      created_at:
        type: string
      id:
        type: integer
      order_num:
        type: string
      product_id:
        type: integer
      status:
        $ref: '#/definitions/model.OrderStatus'
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  model.OrderStatus:
    enum:
    - 0
    - 1
    - 2
    - 3
    type: integer
    x-enum-varnames:
    - OrderStatusUnpaid
    - OrderStatusPaid
    - OrderStatusFailed
    - OrderStatusCancelled
  model.Payment:
    properties:
      amount_cents:
        type: integer
      created_at:
        type: string
      deleted_at:
        type: string
      id:
        type: integer
      notify_data:
        type: string
      order_id:
        type: integer
      payment_id:
        type: string
      status:
        $ref: '#/definitions/model.PaymentStatus'
      updated_at:
        type: string
    type: object
  model.PaymentStatus:
    enum:
    - pending
    - paid
    - failed
    - refunded
    type: string
    x-enum-varnames:
    - PaymentStatusPending
    - PaymentStatusPaid
    - PaymentStatusFailed
    - PaymentStatusRefunded
...
  service.MyCoupon:
    properties:
      amount_cents:
        description: 满减金额（分）
        type: integer
      coupon_id:
        description: 券模板 ID
        type: integer
      description:
        type: string
      discount_rate:
        description: 折扣率，如 90 表示九折
        type: integer
      id:
        description: 用户券 ID
        type: integer
      min_spend_cents:
        description: 使用门槛（分）
        type: integer
      obtained_from:
        description: purchase/vip_month
        type: string
      status:
        $ref: '#/definitions/model.CouponStatus'
      title:
        type: string
      type:
        allOf:
        - $ref: '#/definitions/model.CouponType'
        description: full_cut/discount
      valid_from:
        type: string
      valid_to:
        type: string
    type: object
  service.OrderWithPayment:
    properties:
      coupon:
        $ref: '#/definitions/service.MyCoupon'
      order:
        $ref: '#/definitions/model.Order'
      payment:
        $ref: '#/definitions/model.Payment'
    type: object
...
info:
  contact: {}
  description: SneakerFlash 球鞋秒杀系统接口文档
  title: SneakerFlash API
  version: "1.0"
paths:
...
  /orders:
    get:
      parameters:
      - default: 1
        description: 页码
        in: query
        name: page
        type: integer
      - default: 10
        description: 每页条数
        in: query
        name: page_size
        type: integer
      - description: 订单状态：0未支付 1已支付 2失败 3已取消
        in: query
        name: status
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.OrderListResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 查询订单列表
      tags:
      - 订单
  /stream/orders/{id}:
    get:
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: access token
        in: query
        name: access_token
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream established
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 订阅订单状态推送
      tags:
      - 推送
  /stream/products/{id}:
    get:
      parameters:
      - description: 商品ID
        in: path
        name: id
        required: true
        type: integer
      - description: access token
        in: query
        name: access_token
        type: string
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream established
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 订阅商品库存推送
      tags:
      - 推送
  /orders/{id}:
    get:
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.OrderWithPaymentResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 未找到
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 获取订单详情
      tags:
      - 订单
  /orders/{id}/apply-coupon:
    post:
      consumes:
      - application/json
      parameters:
      - description: 订单ID
        in: path
        name: id
        required: true
        type: integer
      - description: 优惠券参数，coupon_id 为空表示不使用优惠券
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.ApplyCouponReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.OrderWithPaymentResponse'
              type: object
        "400":
          description: 参数错误或状态不允许
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 订单不存在
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 订单应用优惠券
      tags:
      - 订单
  /orders/poll/{order_num}:
    get:
      parameters:
      - description: 订单号
        in: path
        name: order_num
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.PollOrderResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "401":
          description: 未登录
          schema:
            $ref: '#/definitions/app.Response'
      security:
      - BearerAuth: []
      summary: 轮询订单创建状态
      tags:
      - 订单
  /payment/callback:
    post:
      consumes:
      - application/json
      parameters:
      - description: 支付回调参数
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/handler.PaymentCallbackReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/app.Response'
            - properties:
                data:
                  $ref: '#/definitions/handler.OrderWithPaymentResponse'
              type: object
        "400":
          description: 参数错误
          schema:
            $ref: '#/definitions/app.Response'
        "404":
          description: 支付单不存在
          schema:
            $ref: '#/definitions/app.Response'
      summary: 支付回调
      tags:
      - 订单
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"