# internal/handler/swagger_types.go
package handler

import (
	"SneakerFlash/internal/model"
	"SneakerFlash/internal/service"
	"time"
)

// MessageResponse 用于返回简单提示信息。
type MessageResponse struct {
	Message string `json:"message"`
}

// TokenPairResponse 用于登录和刷新 token 响应。
type TokenPairResponse struct {
	AccessToken  string `json:"access_token"`
	RefreshToken string `json:"refresh_token,omitempty"`
	ExpiresIn    int    `json:"expires_in"`
}

// AccessTokenResponse 仅返回新的 access token。
type AccessTokenResponse struct {
	AccessToken string `json:"access_token"`
	ExpiresIn   int    `json:"expires_in"`
}

// IDResponse 返回资源主键。
type IDResponse struct {
	ID int `json:"id"`
}

// OrderNumResponse 返回订单号。
type OrderNumResponse struct {
	OrderNum string `json:"order_num"`
}

// UploadURLResponse 返回上传后资源地址。
type UploadURLResponse struct {
	URL string `json:"url"`
}

// HealthStatusResponse 存活检查响应。
type HealthStatusResponse struct {
	Status    string    `json:"status"`
	Service   string    `json:"service"`
	Timestamp time.Time `json:"timestamp"`
}

// ProbeStatusResponse 单个依赖探针状态。
type ProbeStatusResponse struct {
	Status string `json:"status"`
	Detail string `json:"detail,omitempty"`
}

// ReadinessChecksResponse 就绪检查明细。
type ReadinessChecksResponse struct {
	Database ProbeStatusResponse `json:"database"`
	Redis    ProbeStatusResponse `json:"redis"`
	Kafka    ProbeStatusResponse `json:"kafka"`
}

// ReadinessStatusResponse 就绪检查响应。
type ReadinessStatusResponse struct {
	Status    string                  `json:"status"`
	Service   string                  `json:"service"`
	Timestamp time.Time               `json:"timestamp"`
	Checks    ReadinessChecksResponse `json:"checks"`
}

// UserResponse 用户信息输出。
type UserResponse struct {
	ID              uint      `json:"id"`
	Username        string    `json:"username"`
	Balance         float64   `json:"balance"`
	Avatar          string    `json:"avatar"`
	TotalSpentCents int64     `json:"total_spent_cents"`
	GrowthLevel     int       `json:"growth_level"`
	Role            string    `json:"role"`
	Permissions     []string  `json:"permissions,omitempty"`
	CreatedAt       time.Time `json:"created_at"`
	UpdatedAt       time.Time `json:"updated_at"`
}

// PaymentResponse 用于描述支付单输出，避免暴露内部 gorm.Model。
type PaymentResponse struct {
	ID          uint                `json:"ID"`
	CreatedAt   time.Time           `json:"CreatedAt"`
	UpdatedAt   time.Time           `json:"UpdatedAt"`
	OrderID     uint                `json:"order_id"`
	PaymentID   string              `json:"payment_id"`
	AmountCents int64               `json:"amount_cents"`
	Status      model.PaymentStatus `json:"status"`
	NotifyData  string              `json:"notify_data"`
}

// ProductListResponse 商品列表响应。
type ProductListResponse struct {
	Items []model.Product `json:"items"`
	Total int64           `json:"total"`
	Page  int             `json:"page"`
}

// ProductListWithSizeResponse 包含分页大小的商品列表响应。
type ProductListWithSizeResponse struct {
	Items []model.Product `json:"items"`
	Total int64           `json:"total"`
	Page  int             `json:"page"`
	Size  int             `json:"size"`
}

// OrderListResponse 订单列表响应。
type OrderListResponse struct {
	Items []model.Order `json:"items"`
	Total int64         `json:"total"`
	Page  int           `json:"page"`
	Size  int           `json:"size"`
}

// OrderWithPaymentResponse 订单与支付单组合响应。
type OrderWithPaymentResponse struct {
	Order   *model.Order      `json:"order"`
	Payment *PaymentResponse  `json:"payment,omitempty"`
	Coupon  *service.MyCoupon `json:"coupon,omitempty"`
}

// RiskListResponse 风控名单响应。
type RiskListResponse struct {
	IP   []string `json:"ip"`
	User []string `json:"user"`
}